Extract input style
