+++ src/shared/styles/Common.tsx
